repository_delete_update

+++ product/backend/datamodels/product.go
@@ -3,7 +3,7 @@ package datamodels
 //product模型
 type Product struct {
 	ID           int64  `json:"product_id" sql:"id"`
-	ProductID    int64  `json:"product_id" sql:"product_id"`
+	ProductID    int    `json:"product_id" sql:"product_id"`
 	ProductName  string `json:"product_name" sql:"product_name"`
 	ProductNum   int64  `json:"product_num" sql:"product_num"`
 	ProductImage string `json:"image" sql:"image"`
